+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Hook up Get Group Memberships v2 API [#53](https://github.com/rokwire/rokwire-building-block-sdk-go/issues/53)
+
 ## [1.8.4] - 2025-08-29
 ### Fixed
 - Service token aud validation failing [#47](https://github.com/rokwire/rokwire-building-block-sdk-go/issues/47)